Move parts of OGRLayer dealing with Arrow C interface in a dedicated ogrlayerarrow.cpp file
